requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
joblib
shap
python-pptx